backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.3.0
httpx>=0.26.0

# Google Gemini API
google-generativeai>=0.3.2

# Validation and serialization
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# WebSockets
websockets>=12.0

# Redis for caching/debouncing (optional, falls back to in-memory)
redis>=5.0.1

# Image processing
pillow>=10.0.0

# Async utilities
anyio>=4.2.0

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3

# Development
python-dotenv>=1.0.0
